Support for the CLAP plugin format
This commit adds support for the CLAP plugin format, and I've confirmed
as configured it works in the upcoming Bitwig release. CLAP is an
emerging format found at https://github.com/free-audio/clap

You may not want to merge this as is - I used submodules for the
extensions rather than copy entire libraries as you seemed to do -
but thought you might be interested!